Comprobar en el modelo 3d si las coordenadas de mapeado son correctas

+++ coordinate_xyz.cpp
@@ -17,22 +17,22 @@
 //    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "coordinate_xyz.h"
 
-Coordinate_XYZ::Coordinate_XYZ(QObject *parent,double *coordinate_X,
-                               double *coordinate_Y, double *coordinate_Z) :
+Coordinate_XYZ::Coordinate_XYZ(QObject *parent,float *coordinate_X,
+                               float *coordinate_Y, float *coordinate_Z) :
     QObject(parent)
 {
 }
 //Getters
-double Coordinate_XYZ::getCoordinate_X()
+float Coordinate_XYZ::getCoordinate_X()
 {
  return *_coordinate_X;
 }
 
-double Coordinate_XYZ::getCoordinate_Y()
+float Coordinate_XYZ::getCoordinate_Y()
 {
   return *_coordinate_Y;
 }
-double Coordinate_XYZ::getCoordinate_Z()
+float Coordinate_XYZ::getCoordinate_Z()
 {
  return *_coordinate_Z;
 }
@@ -40,17 +40,17 @@ double Coordinate_XYZ::getCoordinate_Z()
 
 
 //Setters
-void Coordinate_XYZ::setCoordinate_X(double *coordinate_X)
+void Coordinate_XYZ::setCoordinate_X(float *coordinate_X)
 {
 *_coordinate_X= *coordinate_X;
 }
 
-void Coordinate_XYZ::setCoordinate_Y(double *coordinate_Y)
+void Coordinate_XYZ::setCoordinate_Y(float *coordinate_Y)
 {
  *_coordinate_Y= *coordinate_Y;
 }
 
-void Coordinate_XYZ::setCoordinate_Z(double *coordinate_Z)
+void Coordinate_XYZ::setCoordinate_Z(float *coordinate_Z)
 {
  *_coordinate_Z= *coordinate_Z;
 }
